refactor: split bootstrap hook + sandbox helpers per SRP review
Per PR review:
- Extract bootstrap state machine into hooks/useNewChatBootstrap;
  NewChatBootstrap is now a thin renderer.
- Split createSandbox into focused files: parseCreateSandboxErrorResponse,
  getFallbackSandboxCreateErrorMessage, plus lib/string/getOptionalString
  for the generic non-empty-string helper.
- Clarify SandboxCreateRequestError JSDoc — it wraps the chat→recoup-api
  HTTP boundary, not the @vercel/sandbox SDK (which lives server-side
  inside recoup-api).
